requirements.txt
matplotlib==3.10.8
mujoco==3.5.0
numpy==2.4.2
pillow==12.1.1
scipy==1.17.0
